requirements.txt
Flask==3.0.0
Pillow==10.1.0
imagehash==4.3.1
Werkzeug==3.0.1
